Add support for kafka 2.0.0 (#8399)
Add support to kafka 2.0.0 in kafka output and metricbeat module.
Merge kafka versioning helpers of output and metricbeat module.
Set version in kafka module configuration of metricbeat system tests

(cherry picked from commit 1bfd445873eb839ce0c4f3dd353a90ecd0144f8d)
