+++ libbeat/common/kafka/version.go
@@ -23,16 +23,21 @@ import (
 	"github.com/Shopify/sarama"
 )
 
-type Version struct {
-	String string
-}
+// Version is a kafka version
+type Version string
 
+// TODO: remove me.
+// Compat version overwrite for missing versions in sarama
+// Public API is compatible between these versions.
 var (
-	minVersion = sarama.V0_8_2_0
+	v0_10_2_1 = parseKafkaVersion("0.10.2.1")
+	v0_11_0_1 = parseKafkaVersion("0.11.0.1")
+	v0_11_0_2 = parseKafkaVersion("0.11.0.2")
+	v1_0_1    = parseKafkaVersion("1.0.1")
+	v1_0_2    = parseKafkaVersion("1.0.2")
+	v1_1_1    = parseKafkaVersion("1.1.1")
 
 	kafkaVersions = map[string]sarama.KafkaVersion{
-		"": sarama.V0_8_2_0,
-
 		"0.8.2.0": sarama.V0_8_2_0,
 		"0.8.2.1": sarama.V0_8_2_1,
 		"0.8.2.2": sarama.V0_8_2_2,
@@ -49,19 +54,52 @@ var (
 		"0.10.0":   sarama.V0_10_0_1,
 		"0.10.1.0": sarama.V0_10_1_0,
 		"0.10.1":   sarama.V0_10_1_0,
-		"0.10":     sarama.V0_10_1_0,
+		"0.10.2.0": sarama.V0_10_2_0,
+		"0.10.2.1": v0_10_2_1,
+		"0.10.2":   v0_10_2_1,
+		"0.10":     v0_10_2_1,
+
+		"0.11.0.0": sarama.V0_11_0_0,
+		"0.11.0.1": v0_11_0_1,
+		"0.11.0.2": v0_11_0_2,
+		"0.11.0":   v0_11_0_2,
+		"0.11":     v0_11_0_2,
+
+		"1.0.0": sarama.V1_0_0_0,
+		"1.0.1": v1_0_1,
+		"1.0.2": v1_0_2,
+		"1.0":   v1_0_2,
+		"1.1.0": sarama.V1_1_0_0,
+		"1.1.1": v1_1_1,
+		"1.1":   v1_1_1,
+		"1":     v1_1_1,
+
+		"2.0.0": sarama.V2_0_0_0,
+		"2.0":   sarama.V2_0_0_0,
+		"2":     sarama.V2_0_0_0,
 	}
 )
 
-func (v *Version) Validate() error {
-	if _, ok := kafkaVersions[v.String]; !ok {
-		return fmt.Errorf("unknown/unsupported kafka vesion '%v'", v.String)
+func parseKafkaVersion(s string) sarama.KafkaVersion {
+	v, err := sarama.ParseKafkaVersion(s)
+	if err != nil {
+		panic(err)
 	}
+	return v
+}
+
+// Validate that a kafka version is among the possible options
+func (v *Version) Validate() error {
+	if _, ok := kafkaVersions[string(*v)]; !ok {
+		return fmt.Errorf("unknown/unsupported kafka vesion '%v'", *v)
+	}
+
 	return nil
 }
 
+// Unpack a kafka version
 func (v *Version) Unpack(s string) error {
-	tmp := Version{s}
+	tmp := Version(s)
 	if err := tmp.Validate(); err != nil {
 		return err
 	}
@@ -70,10 +108,8 @@ func (v *Version) Unpack(s string) error {
 	return nil
 }
 
-func (v *Version) get() sarama.KafkaVersion {
-	if v, ok := kafkaVersions[v.String]; ok {
-		return v
-	}
-
-	return minVersion
+// Get a sarama kafka version
+func (v Version) Get() (sarama.KafkaVersion, bool) {
+	kv, ok := kafkaVersions[string(v)]
+	return kv, ok
 }
